Backend/requirements.txt
asgiref>=3.6,<4
Django>=4.2,<5.3
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse==0.4.2
python-dotenv==0.19.2
gunicorn>=22.0
dj-database-url==0.5.0
django-ckeditor==6.2.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-smartfields==1.1.3
keyring==5.7.1
pillow==10.4.0
psycopg2-binary>=2.9.9,<3
six==1.16.0
tzdata==2021.5
whitenoise==6.0.0